Refactor FBX parsing types and structure
- Consolidated type imports from ../constants to ../types across multiple parser files.
- Introduced new type definitions for FBX properties, connections, and document structures.
- Created separate files for enums, geometry, materials, textures, and model animations to improve organization.
- Removed obsolete connection node definitions and updated related references.
- Enhanced the ParsingContext class to utilize the new connection types.
- Updated utility functions to align with the new type structure.
- Ensured backward compatibility while streamlining the type system for better maintainability.
