+++ src/types/core/context.ts
@@ -1,7 +1,8 @@
 import type { LoadingManager } from 'three';
 import { Group } from 'three';
 import type { IParsingContext } from './parser';
-import type { IFBXTree, GlobalFBXConnectionNode } from './parser';
+import type { IFBXTree } from '../document';
+import type { FBXConnectionNode } from '../connection';
 
 /**
  * FBX 解析上下文实现
@@ -9,14 +10,14 @@ import type { IFBXTree, GlobalFBXConnectionNode } from './parser';
  */
 export class ParsingContext implements IParsingContext {
   private readonly _fbxTree: IFBXTree;
-  private _connections: Map<number, GlobalFBXConnectionNode>;
+  private _connections: Map<number, FBXConnectionNode>;
   private _sceneGraph: Group;
   private readonly _loadingManager: LoadingManager;
   private readonly _wireframe?: boolean;
 
   constructor (
     fbxTree: IFBXTree,
-    connections: Map<number, GlobalFBXConnectionNode>,
+    connections: Map<number, FBXConnectionNode>,
     loadingManager: LoadingManager,
     wireframe?: boolean
   ) {
@@ -32,11 +33,11 @@ export class ParsingContext implements IParsingContext {
 
   // Getters
   get fbxTree (): IFBXTree { return this._fbxTree; }
-  get connections (): Map<number, GlobalFBXConnectionNode> { return this._connections; }
+  get connections (): Map<number, FBXConnectionNode> { return this._connections; }
   get sceneGraph (): Group { return this._sceneGraph; }
-  
+
   // Setters for properties that need to be modified
-  set connections (value: Map<number, GlobalFBXConnectionNode>) { this._connections = value; }
+  set connections (value: Map<number, FBXConnectionNode>) { this._connections = value; }
   set sceneGraph (value: Group) { this._sceneGraph = value; }
   get loadingManager (): LoadingManager { return this._loadingManager; }
   get wireframe (): boolean | undefined { return this._wireframe; }
@@ -52,7 +53,7 @@ export class ParsingContext implements IParsingContext {
   /**
    * 获取指定 ID 的连接关系
    */
-  getConnections (id: number): GlobalFBXConnectionNode | undefined {
+  getConnections (id: number): FBXConnectionNode | undefined {
     return this._connections.get(id);
   }
 
@@ -77,8 +78,8 @@ export class ParsingContext implements IParsingContext {
   /**
    * 获取与指定节点关联的所有子节点
    */
-  getConnectedNodes (parentId: number): GlobalFBXConnectionNode[] {
-    const connections: GlobalFBXConnectionNode[] = [];
+  getConnectedNodes (parentId: number): FBXConnectionNode[] {
+    const connections: FBXConnectionNode[] = [];
 
     this._connections.forEach(connection => {
       if (connection.parents?.some((parent: { ID: number }) => parent.ID === parentId)) {
@@ -92,7 +93,7 @@ export class ParsingContext implements IParsingContext {
   /**
    * 查找节点的父节点
    */
-  getParentNode (childId: number): GlobalFBXConnectionNode | undefined {
+  getParentNode (childId: number): FBXConnectionNode | undefined {
     return this._connections.get(childId);
   }
 }
