+++ src/types/core/index.ts
@@ -1,3 +1,8 @@
+/**
+ * 核心系统入口
+ * 只使用 export * from
+ */
+
 export * from './parser';
 export * from './context';
 export * from './base-parser';
