+++ src/types/index.ts
@@ -1,61 +1,16 @@
-// 重新导出原有类型，保持向后兼容
-export {
-  IFBXTree,
-  FBXDocumentOptions,
-  FBXProperty,
-  FBXHeaderExtension,
-  FBXConnectionDocment,
-  FBXTimeStamp,
-  FBXSceneInfo,
-  FBXMetaData,
-  FBXTypedProperty,
-  FBXGlobalSettings,
-  FBXLayerElement,
-  FBXLayerElementColor,
-  FBXLayerElementNormal,
-  FBXLayerElementUV,
-  FBXLayer,
-  FBXDeformer,
-  FBXGeometryNode,
-  FBXModelNode,
-  FBXMeshNode,
-  FBXNodeAttribute,
-  FBXAnimationCurveNode,
-  FBXLightNodeAttribute,
-  FBXObjects,
-  FBXMaterial,
-  FBXMaterialNode,
-  FBXTextureNode,
-  FBXVideoNode,
-  FBXPoseNode,
-  FBXSkeleton,
-  RawBone,
-  Deformers,
-  FBXMorphTarget,
-  Global,
-  FBXTree,
-} from '../constants';
+/**
+ * FBX2glTF 类型系统入口
+ * 只使用 export * from
+ */
 
-// 导出节点类型
-export { FBXConnectionNode } from './nodes';
-
-// 导出新的核心类型
-export type {
-  IParsingContext,
-  IParser,
-  IAsyncParser,
-  ModelLoaderResult,
-  ParsingOptions,
-} from './core';
-
-// 导出 ParsingContext 实现
-export { ParsingContext } from './core/context';
-
-// 导出解析器基类
-export { BaseParser, AsyncBaseParser } from './core/base-parser';
-
-// 提供类型别名帮助迁移
-import type { IParsingContext, ModelLoaderResult } from './core';
-
-export type GlobalContext = IParsingContext;
-export type FBXLoaderResult = ModelLoaderResult;
+export * from './enums';
+export * from './base-property';
+export * from './tree-node';
+export * from './connection';
+export * from './layer-element';
+export * from './geometry';
+export * from './material';
+export * from './texture';
+export * from './model-animation';
+export * from './document';
+export * from './core';
