--- src/types/nodes/connection.ts
@@ -1,8 +1,0 @@
-/**
- * FBX 连接节点
- */
-export interface FBXConnectionNode {
-  child: number,
-  parent: number,
-  connectionType: string,
-}
