--- src/types/nodes/index.ts
@@ -1,1 +1,0 @@
-export * from './connection';
